+++ src/server/middleware/session.js
@@ -20,6 +20,6 @@ export default (server) => (req, res, next) => {
 			setSession();
 		}
 	}
-	req.session = new server.Session(sessionId);
+	req.session = server.Session.getInstance(sessionId);
 	next();
 };