+++ src/server/session.js
@@ -5,11 +5,17 @@ import uniqid from 'uniqid';
  */
 export default class Session {
 
+	static _instances = {};
+
 	/**
 	 * Generates the random string as a session id.
 	 */
 	static generateId() {
 		return uniqid();
+	}
+
+	static getInstance(id) {
+		return this._instances[id] || new this(id);
 	}
 
 	/**
@@ -26,7 +32,13 @@ export default class Session {
 	 * @param {string} id Identificator of the session.
 	 */
 	constructor(id) {
-		this.id = id;
+		if (Session._instances[id]) {
+			throw new Error('Session instance already exist.');
+		}
+		if (id) {
+			this.id = id;
+			Session._instances[id] = this;
+		}
 	}
 
 	/**